requirements.txt
numpy
pillow
tifffile
matplotlib
Jinja2
